refactor(client): 移动SeismicBounceManager到support包
- 将SeismicBounceManager类从renderer.shake包重构至client.support包
- 更新所有引用SeismicBounceManager的导入路径
- 保持SeismicBounceManager功能逻辑不变，确保兼容性
