+++ src/main/java/dev/dubhe/anvilcraft/network/GiantAnvilShockEffectPacket.java
@@ -3,7 +3,7 @@ package dev.dubhe.anvilcraft.network;
 import dev.anvilcraft.lib.v2.network.packet.IClientboundPacket;
 import dev.anvilcraft.lib.v2.network.packet.IPacket;
 import dev.dubhe.anvilcraft.AnvilCraft;
-import dev.dubhe.anvilcraft.client.renderer.shake.SeismicBounceManager;
+import dev.dubhe.anvilcraft.client.support.SeismicBounceManager;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.codec.ByteBufCodecs;
